发布script(1.1.6)： 修改记录：     一、lib脚本修改 	lib脚本版本号：1.1.6 	(1)linkssl：修正“存在ssl socket连接的情况下，再新建一个ssl连接，新连接配置证书失败”的问题 		【问题重要性】：一般 		【问题影响模块】：Air202、Air800、Air801、Air201、Air208 		【问题出现概率】：必现 		【问题出现原因】：存在ssl连接的情况下，有一个初始化标志设置为TRUE，此标志会导致后续的新连接证书无法执行导入的程序 		【问题解决方式】：证书导入功能和初始化标志不再关联 	(2)aliyuniotota：新增“阿里云OTA（设备固件升级）”的功能 		【新功能重要性】：一般 		【新功能影响模块】：Air202、Air800、Air801、Air201、Air208 		【新功能说明】：升级包放在阿里云iot控制台上，参考aliyun_ssl_huadong2的demo说明     二、demo脚本修改 	(1)aliyun_ssl_huadong2：新增阿里云OTA功能的使用说明 	(2)proto_buffer：新增"google proto buffer编解码测试demo"，需要配合大于等于0020版本的lod使用 	(3)crypto：新增"计算文件的md5值"的使用说明，需要配合大于等于0020版本的lod使用 	(4)default：修改功能，每秒一次，uart1和uart2向外输出“烧写的软件为Luat软件，不是AT软件”的提示 	(5)http、https：修正“http和https的demo中，获取的数据保存到文件中时，如果保存失败，脚本重启”的问题 	(6)crypto：新增"aes.encrypt和aes.decrypt"接口的使用说明，需要配合大于等于0020版本的lod使用

+++ script/demo/http/test.lua
@@ -82,9 +82,9 @@ local function rcvcbfile(result,statuscode,head,filename)
 		end
 		
 		filehandle:close()
-		--文件使用完之后，如果以后不再需求，需要自行删除
-		os.remove(filename)
 	end
+	--文件使用完之后，如果以后不再需求，需要自行删除
+	if filename then os.remove(filename) end
 	
 	httpclient:disconnect(discb)
 end
